+++ src/Skeletest/Plugin.hs
@@ -1,4 +1,5 @@
 module Skeletest.Plugin (
+  -- * Plugin
   Plugin (..),
   defaultPlugin,
 
@@ -22,13 +23,19 @@ module Skeletest.Plugin (
   hasMarkerNamed,
 ) where
 
+import Control.Monad ((>=>))
 import Skeletest.Internal.CLI (Flag)
 import Skeletest.Internal.Markers (findMarker, hasMarkerNamed)
 import Skeletest.Internal.Snapshot (SnapshotRenderer)
 import Skeletest.Internal.Spec.Output (BoxSpec, BoxSpecContent (..))
+import Skeletest.Internal.Spec.Tree (SpecTree)
 import Skeletest.Internal.TestInfo (TestInfo (..))
 import Skeletest.Internal.TestRunner (TestResult (..), TestResultMessage (..))
 
+-- | A plugin for extending Skeletest.
+--
+-- Use 'defaultPlugin' instead of using v'Plugin' directly, to minimize
+-- breaking changes.
 data Plugin = Plugin
   { cliFlags :: [Flag]
   , snapshotRenderers :: [SnapshotRenderer]
@@ -54,14 +61,22 @@ defaultPlugin =
     , hooks = defaultHooks
     }
 
+-- | Hooks for extending Skeletest.
+--
+-- Use 'defaultHooks' instead of using v'Hooks' directly, to minimize
+-- breaking changes.
 data Hooks = Hooks
-  { hookRunTest :: TestInfo -> IO TestResult -> IO TestResult
+  { hookModifyFileSpecs :: [SpecTree] -> IO [SpecTree]
+  -- ^ Modify the specs in a file
+  , hookRunTest :: TestInfo -> IO TestResult -> IO TestResult
+  -- ^ Modify how a test is executed
   }
 
 instance Semigroup Hooks where
   hooks1 <> hooks2 =
     Hooks
-      { hookRunTest = \testInfo -> hookRunTest hooks2 testInfo . hookRunTest hooks1 testInfo
+      { hookModifyFileSpecs = hookModifyFileSpecs hooks1 >=> hookModifyFileSpecs hooks2
+      , hookRunTest = \testInfo -> hookRunTest hooks2 testInfo . hookRunTest hooks1 testInfo
       }
 
 instance Monoid Hooks where
@@ -70,5 +85,6 @@ instance Monoid Hooks where
 defaultHooks :: Hooks
 defaultHooks =
   Hooks
-    { hookRunTest = \_ -> id
+    { hookModifyFileSpecs = pure
+    , hookRunTest = \_ -> id
     }